python_core_lib/pyproject.toml
[tool.poetry]
name = "python_core_lib"
version = "0.1.0"
description = "Core Python Utilities Library"
authors = ["Zachi Nachshon <zachi.nachshon@gmail.com>"]
exclude = [
    "**/*_test*", 
    "**/*_fake*", 
    "dist", 
    "htmlcov", 
    "poetry.lock", 
    "pytest", 
    "Makefile", 
    "runners", 
    ".gitignore",
    ".coverage",
    ".coveragerc",
    ".git-deps",
    ".pytest_cache"]

packages = [
    { include = "python_core_lib/runner/ansible/resources" },
    { include = "python_core_lib/runner/ansible/resources/callback_plugins" }
]

include = [
    { path = "python_core_lib/runner/ansible/resources/ansible.cfg" },
]

[tool.black]
line-length = 120

[tool.ruff]
exclude = [".venv", "external", ".git-deps"]
line-length = 120
ignore = [ "E501", "E402", "F403", "F405" ]
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    # isort
    "I",
    # Whitespaces
    "W", 
]

[tool.mypy]
ignore_missing_imports = "True"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.6.1"
loguru = "^0.6.0"
requests = "^2.28.1"
rich = "^12.5.1"      # - Progress Bar for Python and CLI (https://github.com/Textualize/rich)
python3-nmap = "^1.5.4"
python-libnmap = "^0.7.3"
PyYAML = "^6.0"
inquirer = "^2.10.0"
python-hosts = "^1.0.3"
ansible = "^7.3.0"
ansible-runner = "^2.3.2"
# GitPython = "^3.1.18" # - GitPython provides object model access to your git repository (https://github.com/gitpython-developers/GitPython)
# #Jinja2 = "^3.0.3"      - Templating engine (https://github.com/pallets/jinja)
# #sh = "^1.14.2"         - Subprocess replacement for Python (https://github.com/amoffat/sh)
# #backoff = "^1.10.0"    - Function decoration for backoff and retry (https://github.com/litl/backoff)
# #tenacity = "^8.0.1"    - General-purpose retrying library (https://github.com/jd/tenacity)

[tool.poetry.dev-dependencies]
black = "^22.8.0"
pytest = "^7.1.3"
mypy = "^0.971"
coverage = "^6.4.4"
ruff = "^0.0.265"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"